requirements.txt
annotated-types==0.6.0
asgiref==3.8.0
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.5
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
packaging==24.0
POT==0.9.3
preshed==3.0.9
pydantic==2.6.4
pydantic_core==2.16.3
python-vlc==3.0.20123
regex==2023.12.25
requests==2.31.0
setuptools==69.2.0
smart-open==6.4.0
sqlparse==0.4.4
srsly==2.4.8
thinc==8.2.3
threadpoolctl==3.4.0
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
wheel==0.43.0
zeroc-ice==3.7.10
